Use integration tests for debug info.
